requirements.txt
# DND.AI dependencies
discord.py[voice]>=2.3.0
discord-ext-voice-recv>=0.3.0
python-dotenv>=1.0.0
pynacl>=1.5.0
